pyproject.toml
[tool.poetry]
name = "tcc"
version = "0.1.0"
description = ""
authors = ["Gustavo Moser <gustavovbmoser@gmail.com>"]
packages = [
    { include = "articleservice", from = "src" },
    { include = "db", from = "src" },
    { include = "service", from = "src" },
    { include = "twitterservice", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.10.1"
Scrapy = "^2.6.2"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
emoji-translate = "^0.1.1"
regex = "^2022.9.13"
spacy = "^3.4.1"

[tool.poetry.scripts]
save_articles = "src.articles:saveArticles"
save_tweets = "src.tweets:saveTweets"
clean_tweets = "src.tweets:cleanTweets"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
